requirements.txt
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==19.1.0
asgiref==3.5.0
astroid==2.5.6
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
black==21.9b0
boto3==1.21.18
botocore==1.24.18
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
Collectfast==1.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==36.0.1
decorator==5.0.9
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.49.0
django-anymail==8.5
django-coverage-plugin==1.6.0
django-crispy-forms==1.8.1
django-environ==0.4.5
django-extensions==2.2.5
django-model-utils==3.2.0
django-pipeline==1.7.0
django-redis==4.11.0
django-storages==1.12.3
djangorestframework==3.11.2
docutils==0.17.1
entrypoints==0.3
factory-boy==2.12.0
Faker==2.0.1
flake8==3.7.8
gunicorn==20.1.0
idna==2.10
imagesize==1.2.0
ipdb==0.12.2
ipython==7.31.1
ipython-genutils==0.2.0
isort==5.8.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
lazy-object-proxy==1.6.0
Markdown==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
more-itertools==8.8.0
mypy==0.720
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==20.9
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.0
pluggy==0.13.1
progress==1.5
prompt-toolkit==3.0.18
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.9.0
PyJWT==2.3.0
pylint==2.8.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.1.1
pytest-django==3.5.1
pytest-sugar==0.9.2
python-dateutil==2.8.1
python-slugify==3.0.3
python3-openid==3.2.0
pytz==2019.2
redis==3.4.1
regex==2021.10.8
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.2
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stripe==2.40.0
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.2
tomli==1.2.1
traitlets==4.3.3
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==0.15.3
wrapt==1.12.1
